refactor: chessGame 패키지 이동, 줄 정리
