+++ src/main/java/chess/domain/chessGame/ChessGame.java
@@ -1,11 +1,11 @@
-package chess;
+package chess.domain.chessGame;
 
 import chess.domain.position.Position;
 
 import java.util.Map;
 
 public interface ChessGame {
-    Map<Position, String> move(String currentPosition, String nextPosition); // move b1 b2
+    Map<Position, String> move(String currentPosition, String nextPosition);
     ChessGame start();
     void end();
     boolean isEnd();
